Latest zf2

+++ tests/ZendTest/Form/View/Helper/Captcha/ImageTest.php
@@ -3,9 +3,8 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2013 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
- * @package   Zend_Form
  */
 
 namespace ZendTest\Form\View\Helper\Captcha;
@@ -16,14 +15,10 @@ use Zend\Form\Element\Captcha as CaptchaElement;
 use Zend\Form\View\Helper\Captcha\Image as ImageCaptchaHelper;
 use ZendTest\Form\View\Helper\CommonTestCase;
 
-/**
- * @category   Zend
- * @package    Zend_Form
- * @subpackage UnitTest
- */
 class ImageTest extends CommonTestCase
 {
     protected $tmpDir;
+    protected $testDir;
 
     public function setUp()
     {
@@ -62,11 +57,17 @@ class ImageTest extends CommonTestCase
     public function tearDown()
     {
         // remove captcha images
-        foreach (new DirectoryIterator($this->tmpDir) as $file) {
+        if (!isset($this->testDir)) {
+            parent::tearDown();
+            return;
+        }
+
+        foreach (new DirectoryIterator($this->testDir) as $file) {
             if (!$file->isDot() && !$file->isDir()) {
                 unlink($file->getPathname());
             }
         }
+        parent::tearDown();
     }
 
     /**
