+++ tests/ZendTest/Mail/Storage/MaildirFolderTest.php
@@ -3,9 +3,8 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2013 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
- * @package   Zend_Mail
  */
 
 namespace ZendTest\Mail\Storage;
@@ -14,9 +13,6 @@ use Zend\Config;
 use Zend\Mail\Storage\Folder;
 
 /**
- * @category   Zend
- * @package    Zend_Mail
- * @subpackage UnitTests
  * @group      Zend_Mail
  */
 class MaildirFolderTest extends \PHPUnit_Framework_TestCase
@@ -89,6 +85,9 @@ class MaildirFolderTest extends \PHPUnit_Framework_TestCase
                 if (!file_exists($this->_tmpdir . $dir . '/' . $subdir)) {
                     continue;
                 }
+                if (!is_dir($this->_tmpdir . $dir . '/' . $subdir)) {
+                    continue;
+                }
                 $dh = opendir($this->_tmpdir . $dir . '/' . $subdir);
                 while (($entry = readdir($dh)) !== false) {
                     $entry = $this->_tmpdir . $dir . '/' . $subdir . '/' . $entry;
@@ -100,7 +99,7 @@ class MaildirFolderTest extends \PHPUnit_Framework_TestCase
                 closedir($dh);
                 rmdir($this->_tmpdir . $dir . '/' . $subdir);
             }
-            if ($dir != '.') {
+            if ($dir != '.' && is_dir($this->_tmpdir . $dir)) {
                 rmdir($this->_tmpdir . $dir);
             }
         }
